% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedata.R
\docType{data}
\name{expedata}
\alias{expedata}
\title{Data from FMS instrument}
\format{
\subsection{\code{expedata}}{

A data frame with 7799 rows and 9 columns. Important columns described below.
\describe{
\item{Time}{Time in hh:mm:ss}
\item{O2}{Oxygen levels}
\item{CO2}{Carbon dioxide levels}
\item{FR}{Flow rate}
...
}
}
}
\usage{
expedata
}
\description{
A dataframe of FMS measures, as exported by expedata.
}
\keyword{datasets}
