% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedata.R, R/markers.R
\docType{data}
\name{markers}
\alias{markers}
\title{Data from FMS instrument}
\format{
\subsection{\code{expedata}}{

A data frame with 7799 rows and 9 columns. Important columns described below.
\describe{
\item{Time}{Time in hh:mm:ss}
\item{O2}{Oxygen levels}
\item{CO2}{Carbon dioxide levels}
\item{FR}{Flow rate}
...
}
}

\subsection{\code{markers}}{

A data frame with 11 rows and 4 columns:
\describe{
\item{second}{Seconds from the beginning of the run}
\item{marker}{Label of what the timepoint is; should be B (baseline), S (sample), or end (end of these sample periods, or to exclude noisy sections).}
\item{number}{Numeric order of events; indicates the order of baseline and sample periods for slope calculations}
...
}
}
}
\usage{
markers

markers
}
\description{
A dataframe of FMS measures, as exported by expedata.

A dataframe of timestamps for labelling incurrent and excurrent sections of respirometry data. Importantly, use "end" to exclude noise in the data.
}
\keyword{datasets}
