% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VO2FMS.R
\name{VO2FMS}
\alias{VO2FMS}
\title{Calculate Respirometry Metrics}
\usage{
VO2FMS(data, markers)
}
\arguments{
\item{data}{Exported expedata file}

\item{markers}{Dataframe of timepoints and labels}
}
\value{
A named list containing:
\describe{
\item{data}{The full joined dataset with VO2, VCO2, and rolling means.}
\item{VO2_min_candidates}{Top 10 rows with lowest 30s VO2.}
\item{VO2_max_candidates}{Top 10 rows with highest 30s VO2.}
\item{VCO2_min_candidates}{Top 10 rows with lowest 30s VCO2.}
\item{VCO2_max_candidates}{Top 10 rows with highest 30s VCO2.}
}
}
\description{
Calculate Respirometry Metrics
}
