% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEdriftR.R
\name{TEdriftR}
\alias{TEdriftR}
\title{Get thermal endurance metrics from LabChart output}
\usage{
TEdriftR(data)
}
\arguments{
\item{data}{A dataframe.}
}
\value{
A list of the original data, and dataframes of 10 candidates for important values.
}
\description{
Get thermal endurance metrics from LabChart output
}
