% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VO2FMS.R
\name{VO2FMS}
\alias{VO2FMS}
\title{Get thermal endurance metrics from FMS output}
\usage{
VO2FMS(data, markers)
}
\arguments{
\item{data}{A dataframe with 4+ columns. Four of the columns must have the names: O2 = oxygen levels, CO2 = CO2 levels, FR = the flow rate, Time = current time in hh:mm:ss.}

\item{markers}{A dataframe with 2 columns, second and marker. Second is the time since the start of the trial, in seconds. Marker is "B" for baseline (incurrent) or "S" for sample (Excurrent)}
}
\value{
A list of the original data, and dataframes of 10 candidates for important values.
}
\description{
This function calculates respirometry metrics from a single-analyzer setup, that accounts for drift over time. This works by taking incurrent measurements throughout the trialling period to create a 'drift slope' throughout the run, which corrects excurrent measures. These corrected measures are then used to calculate rolling VO2 max measures.
}
\details{
The \verb{data$Cmt Text} variable \strong{must} have the following levels.
"Baseline.Incur.Start", "Baseline.Excur.Start", "Last.Incur.Start", "Last.Excur.Start". These are baseline measurements taken with no speciemen in the chamber.
"Incur.X.Start" are measurements taken from the incurrent (before speciement chamber). X is a placeholder for numbers 1:5, indicating the sections during trial.
"Excur.X.Start" are measurments taken from the excurrent (afte speciemen chamber). X is a placeholder for numbers 1:5, indicating the sections during trial.
}
